Upgrade to sbt-scalajs 1.0.0-RC2 (used only by `client`).
This means that the `client` project is now built with Scala.js
1.0.0-RC2. However, the snippets and the `page` module are still
compiled and linked by Scala.js 0.6.31.

At this stage, we have a proof of concept that the version of
Scala.js used to build the `client` project is completely
independent from the version exposed to users of ScalaFiddle.
This paves the way to expose *several* versions of Scala.js in the
future.